requirements.txt
wheel==0.42.0
Mastodon.py==1.8.1
schedule==1.2.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
feedparser==6.0.10
fake-useragent==1.1.1
facebook-scraper==0.2.59
urllib3<2.0.0
sgmllib3k==1.0.0
demjson3==3.0.5
requests-html==0.10.0
pyppeteer==1.0.2
lxml[html_clean]==4.9.3